refactor: replace fiat formatter with payFiatFormatter for transaction details and centralize user currency handling
